Cleanup and document most of `cosy-macro`
