+++ cosy-macro/src/lib.rs
@@ -1,5 +1,10 @@
+//! Cosy reactor proc macro.
+//!
+//! Generates reactors with routing from concise definitions. See `README.md` for details.
+
 mod gen;
 mod parse;
+mod rust_type;
 mod util;
 
 use proc_macro::TokenStream;
@@ -7,6 +12,7 @@ use syn::parse_macro_input;
 
 use parse::ReactorDefinition;
 
+/// Generates a new reactor implementation, along with types.
 #[proc_macro]
 pub fn reactor(input: TokenStream) -> TokenStream {
     let def = parse_macro_input!(input as ReactorDefinition);
